backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
transformers>=4.48.0
# CPU-only PyTorch (no CUDA required)
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.2.0
numpy>=1.26.0
requests==2.31.0
python-dateutil==2.8.2
yfinance>=0.2.0
pandas>=2.0.0
requests-cache>=1.0.0
APScheduler>=3.10.0
# google-genai not needed: agent uses Gemini REST API via requests
